requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
boto3==1.12.5
botocore==1.15.5
Django==3.0.3
django-cors-headers==3.2.1
django-extensions==2.2.8
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-storages==1.8
docutils==0.15.2
graphene==2.1.8
graphene-django==2.8.2
graphene-file-upload==1.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
jmespath==0.9.4
Pillow==7.0.0
promise==2.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
